+++ AlchemistCode/Cards/Uncommon/Overspill.cs
@@ -1,0 +1,24 @@
+using Alchemist.AlchemistCode.Commands;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+
+namespace Alchemist.AlchemistCode.Cards.Uncommon;
+
+[CardTheme(CardTheme.Mix)]
+public class Overspill : AlchemistCard
+{
+    public Overspill() : base(2, CardType.Attack, CardRarity.Uncommon, TargetType.AllEnemies)
+    {
+        WithDamage(8, 2);
+        WithTips(card => Mixing.MixTips(card.IsUpgraded));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_slime_impact"))
+            .WithAttackerAnim("Cast", Owner.Character.CastAnimDelay)
+            .Execute(choiceContext);
+        await Mixing.CreateRandom(choiceContext, Owner, IsUpgraded);
+    }
+}